[v1.15] backport: fix google.protobuf.UInt64Value does not marshal correctly in CRDs (#5303)
* bump gloo

* make generate-all

* Fix type in cache plugin

+++ projects/gloo/pkg/plugins/caching/plugin.go
@@ -90,7 +90,7 @@ func (p *plugin) HttpFilters(params plugins.Params, listener *v1.HttpListener) (
 
 	grpcConfig := &grpc.GrpcCacheConfig{
 		Service:        svc,
-		MaxPayloadSize: maxSize, // set the max size as enforced by the grpc cache
+		MaxPayloadSize: uint64(maxSize), // set the max size as enforced by the grpc cache
 	}
 
 	marshalledGRPCConfig, err := utils.MessageToAny(grpcConfig)
